backend/requirements.txt
Flask==2.3.3
requests==2.31.0
openai==1.3.5
python-dotenv==1.0.1
